Organize our components, pages, and images
Following on the pattern established in
50679e689491e39cb6090b86275dd3198219cb6d, organize our pages and
components into their respective directories, and also move images into
a new `/src/images` directory.

The entry-point (index route and and `App` component) remain at the root
of `/src` to reflect their primacy.

Issue #10 Scaffold routing and page construction
